CI: Add ACVP tests in CI
This commit adds a portable acvp_client.py that runs all the ACVP tests in
parallel. This way we do not have to rely on parallel to be installed.
It also adds these ACVP tests to CI.

Resolves https://github.com/pq-code-package/slhdsa-c/issues/4
