+++ .github/workflows/all.yml
@@ -13,6 +13,13 @@ on:
     branches: ["main"]
     types: [ "opened", "synchronize" ]
 jobs:
+  base:
+    name: Base
+    permissions:
+      contents: 'read'
+      id-token: 'write'
+    uses: ./.github/workflows/base.yml
+    secrets: inherit
   nix:
     name: Nix
     permissions:
@@ -26,7 +33,6 @@ jobs:
     permissions:
       contents: 'read'
       id-token: 'write'
-    needs: [ nix ]
+    needs: [ base, nix ]
     uses: ./.github/workflows/cbmc.yml
     secrets: inherit
-
